Add: knot_message_item
When the thing acks the config or the GW sends a get_data they need
to send only the sensor_id. This new message is used to indicate from
which item (sensor/actuator) the data was sent, and the operation
performed is based on the header type. It makes easier to the
developers understand the code and reduce the payload for some message
types.